+++ plugins/websocket_server/WebsocketServer.cc
@@ -19,6 +19,7 @@
 #include <ignition/common/Util.hh>
 #include <ignition/msgs.hh>
 
+#include "MessageDefinitions.hh"
 #include "WebsocketServer.hh"
 
 using namespace ignition::launch;
@@ -268,10 +269,19 @@ void WebsocketServer::OnDisconnect(int _socketId)
 //////////////////////////////////////////////////
 void WebsocketServer::OnMessage(int _socketId, const std::string &_msg)
 {
+  // Handle the case where the client requests the message definitions.
+  if (_msg == "message_definitions")
+  {
+    igndbg << "Message definitions request recieved\n";
+    this->QueueMessage(this->connections[_socketId].get(),
+        kMessageDefinitions.c_str(), kMessageDefinitions.length());
+    return;
+  }
+
   ignition::msgs::WebRequest requestMsg;
   requestMsg.ParseFromString(_msg);
 
-  if (requestMsg.operation() == "list")
+  if (requestMsg.operation() == "topic_list")
   {
     igndbg << "Topic list request recieved\n";
     ignition::msgs::Packet msg;
@@ -285,6 +295,8 @@ void WebsocketServer::OnMessage(int _socketId, const std::string &_msg)
     for (const std::string &topic : topics)
       msg.mutable_string_msg_v()->add_data(topic);
 
+    msg.set_topic("/topic_list");
+    msg.set_type("ignition.msgs.StringMsg_V");
     std::string data = msg.SerializeAsString();
 
     // Queue the message for delivery.