+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ### Ignition Launch 1.X.X
 
+1. Added ability to get protobuf definitions from the websocket server
+    * [Pull Request 21](https://github.com/ignitionrobotics/ign-launch/pull/21)
+
 1. Added ability to set the publication rate of the websocket server.
     * [Pull Request 17](https://github.com/ignitionrobotics/ign-launch/pull/17)